requirements.txt
fastapi
python-dotenv
uvicorn
python-socketio
requests
beautifulsoup4
googlesearch-python
together
langchain
langchain-community
langchain-core
langchain-together
faiss-cpu
sentence-transformers
docx2txt
python-docx
pandas
openpyxl
matplotlib
plotly
fpdf2
PyJWT
python-jose[cryptography]
Pillow
kaleido
SQLAlchemy
pydantic
pypdf
unstructured[all-docs]
layoutparser
pytesseract
langdetect
PyMuPDF
pdfplumber
python-multipart
pytz
